Clones added
After a lot of sweat and tears, I finally managed to do this. It
required me to place a list of clones on e2dElement, so that whenever
some method moves an element, it can apply the inverse transformation to
all the clones, so that the clones remain the same.